adjust old test code
