+++ tests/Unit/inc/Engine/Cache/PurgeExpired/PurgeExpiredCache/purgeExpiredFiles.php
@@ -1,48 +1,50 @@
 <?php
 
-namespace WP_Rocket\Tests\Integration\inc\classes\Cache\Expired_Cache_Purge;
+namespace WP_Rocket\Tests\Unit\inc\Engine\Cache\PurgeExpired\PurgeExpiredCache;
 
+use Brain\Monkey\Filters;
 use Brain\Monkey\Functions;
-use WP_Rocket\Cache\Expired_Cache_Purge;
-use WP_Rocket\Tests\Integration\FilesystemTestCase;
+use WP_Rocket\Engine\Cache\PurgeExpired\PurgeExpiredCache;
+use WP_Rocket\Tests\Unit\FilesystemTestCase;
 
 /**
- * @covers \WP_Rocket\Cache\Expired_Cache_Purge::purge_expired_files
- * @uses   ::get_rocket_i18n_uri
- * @uses   ::rocket_direct_filesystem
- * @uses   ::get_rocket_parse_url
+ * @covers \WP_Rocket\Engine\Cache\PurgeExpired\PurgeExpiredCache::purge_expired_files
  * @uses   \WP_Rocket\Buffer\Cache::can_generate_caching_files
  * @group  Cache
  * @group  vfs
  */
 class Test_PurgeExpiredFiles extends FilesystemTestCase {
-	protected $path_to_test_data = '/inc/classes/Cache/Expired_Cache_Purge/purgeExpiredFiles.php';
+	protected $path_to_test_data = '/inc/Engine/Cache/PurgeExpired/PurgeExpiredCache/purgeExpiredFiles.php';
+
+	public static function setUpBeforeClass() {
+		parent::setUpBeforeClass();
+
+		require_once WP_ROCKET_PLUGIN_ROOT . 'inc/functions/i18n.php';
+	}
 
 	public function testShouldReturnNullWhenNoLifespan() {
-		Functions\expect( 'get_rocket_il8n_uri' )->never();
+		Functions\expect( 'get_rocket_i18n_uri' )->never();
 		Functions\expect( 'rocket_direct_filesystem' )->never();
 
-		$expired_cache_purge = new Expired_Cache_Purge( $this->filesystem->getUrl( 'wp-content/cache/wp-rocket' ) );
+		$expired_cache_purge = new PurgeExpiredCache( '' );
 		$this->assertNull( $expired_cache_purge->purge_expired_files( 0 ) );
+		$this->assertEquals( 0, Filters\applied( 'rocket_automatic_cache_purge_urls' ) );
 	}
 
 	public function testShouldReturnNullWhenNoURLs() {
+		Functions\expect( 'get_rocket_i18n_uri' )
+			->once()
+			->andReturn( [ null, 1, '' ] );
 		Functions\expect( 'rocket_direct_filesystem' )->never();
 
-		add_filter( 'rocket_automatic_cache_purge_urls', '__return_empty_array' );
-
-		$this->setFilesToExpire( $this->original_files );
-
-		$expired_cache_purge = new Expired_Cache_Purge( $this->filesystem->getUrl( 'wp-content/cache/wp-rocket' ) );
+		$expired_cache_purge = new PurgeExpiredCache( '' );
 		$this->assertNull( $expired_cache_purge->purge_expired_files( 36000 ) );
+		$this->assertEquals( 1, Filters\applied( 'rocket_automatic_cache_purge_urls' ) );
 
 		// Check that no files were purged, i.e. just to make sure.
 		foreach ( $this->original_files as $file ) {
 			$this->assertTrue( $this->filesystem->exists( $file ) );
 		}
-
-		// Clean up.
-		remove_filter( 'rocket_automatic_cache_purge_urls', '__return_empty_array' );
 	}
 
 	/**
@@ -56,10 +58,21 @@ class Test_PurgeExpiredFiles extends FilesystemTestCase {
 			$this->assertTrue( $this->filesystem->exists( $file ) );
 		}
 
+		Functions\expect( 'get_rocket_i18n_uri' )->once()->andReturn( [ 'http://example.org/' ] );
+		Functions\expect( 'get_rocket_parse_url' )
+			->once()
+			->andReturnUsing(
+				function( $value ) {
+					return parse_url( $value );
+				}
+			);
+
 		// Purge the expired files.
-		$expired_cache_purge = new Expired_Cache_Purge( $this->filesystem->getUrl( $this->config['vfs_dir'] ) );
+		$expired_cache_purge = new PurgeExpiredCache( $this->filesystem->getUrl( $this->config['vfs_dir'] ) );
 		$expired_cache_purge->purge_expired_files( $lifespan );
 
+		$this->assertEquals( 1, Filters\applied( 'rocket_automatic_cache_purge_urls' ) );
+		$this->assertEquals( 1, Filters\applied( 'rocket_url_no_dots' ) );
 		$this->assertEquals( 1, did_action( 'rocket_before_automatic_cache_purge_dir' ) );
 		$this->assertEquals( 1, did_action( 'rocket_after_automatic_cache_purge_dir' ) );
 		$this->assertEquals( 1, did_action( 'rocket_after_automatic_cache_purge' ) );
